requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.8
docker==7.1.0
filelock==3.16.1
frozenlist==1.4.1
idna==3.10
joblib==1.4.2
magic-filter==1.0.12
multidict==6.1.0
nltk==3.9.1
platformdirs==4.3.6
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core>=2.0,<3.0
python-decouple==3.8
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.26.6
yarl==1.14.0
python-dotenv==0.21.0
pandas==2.2.3
